Changed Piece Shapes and Dropping Mechanics
To try and make the buildings less 'straggly' I've rotated the 'L' block to lie flat.

In addition, I've added a mechanism where when a block is added, it picks the 'best' column to minimise the height of the tower. In other words, it looks for the column where it can fall the most. If there are multiple 'best' choices, it picks one randomly.